Collapse the OOB framework
We played with alternative implementations over the years,
but nothing ever really stuck. So let's simplify the code
by removing the framework and the associated loopbacks
that were built into it (e.g., if a message cannot be sent,
then loop it back into the OOB base to see if another
component is available that can send it).

The code badly needs reorganization as I've made no
attempt to do so here. A pass to see if event steps
can be eliminated would also be good - I've cleaned
up a few of them, but what remains could use another
pair of eyes.
