--- src/mca/oob/Makefile.am
@@ -1,40 +1,0 @@
-#
-# Copyright (c) 2004-2005 The Trustees of Indiana University and Indiana
-#                         University Research and Technology
-#                         Corporation.  All rights reserved.
-# Copyright (c) 2004-2005 The University of Tennessee and The University
-#                         of Tennessee Research Foundation.  All rights
-#                         reserved.
-# Copyright (c) 2004-2005 High Performance Computing Center Stuttgart,
-#                         University of Stuttgart.  All rights reserved.
-# Copyright (c) 2004-2005 The Regents of the University of California.
-#                         All rights reserved.
-# Copyright (c) 2010-2020 Cisco Systems, Inc.  All rights reserved
-# Copyright (c) 2019      Intel, Inc.  All rights reserved.
-# Copyright (c) 2022      Nanook Consulting.  All rights reserved.
-# $COPYRIGHT$
-#
-# Additional copyrights may follow
-#
-# $HEADER$
-#
-
-# main library setup
-noinst_LTLIBRARIES = libprtemca_oob.la
-libprtemca_oob_la_SOURCES =
-
-# pkgdata setup
-dist_prtedata_DATA =
-
-# local files
-headers = oob.h
-libprtemca_oob_la_SOURCES += $(headers)
-
-# Conditionally install the header files
-prtedir = $(prteincludedir)/$(subdir)
-nobase_prte_HEADERS = $(headers)
-
-include base/Makefile.am
-
-distclean-local:
-	rm -f base/static-components.h
