--- src/mca/oob/base/Makefile.am
@@ -1,32 +1,0 @@
-#
-# Copyright (c) 2004-2005 The Trustees of Indiana University and Indiana
-#                         University Research and Technology
-#                         Corporation.  All rights reserved.
-# Copyright (c) 2004-2005 The University of Tennessee and The University
-#                         of Tennessee Research Foundation.  All rights
-#                         reserved.
-# Copyright (c) 2004-2005 High Performance Computing Center Stuttgart,
-#                         University of Stuttgart.  All rights reserved.
-# Copyright (c) 2004-2005 The Regents of the University of California.
-#                         All rights reserved.
-# Copyright (c) 2012-2013 Los Alamos National Security, LLC. All rights
-#                         reserved.
-# Copyright (c) 2014-2019 Intel, Inc.  All rights reserved.
-# Copyright (c) 2020      Cisco Systems, Inc.  All rights reserved
-# Copyright (c) 2022      Nanook Consulting.  All rights reserved.
-# $COPYRIGHT$
-#
-# Additional copyrights may follow
-#
-# $HEADER$
-#
-
-dist_prtedata_DATA += base/help-oob-base.txt
-
-headers += \
-	base/base.h
-
-libprtemca_oob_la_SOURCES += \
-        base/oob_base_stubs.c \
-        base/oob_base_frame.c \
-        base/oob_base_select.c
