--- src/mca/oob/base/owner.txt
@@ -1,7 +1,0 @@
-#
-# owner/status file
-# owner: institution that is responsible for this package
-# status: e.g. active, maintenance, unmaintained
-#
-owner: project
-status: maintenance
