--- src/mca/oob/tcp/configure.m4
@@ -1,39 +1,0 @@
-# -*- shell-script -*-
-#
-# Copyright (c) 2004-2005 The Trustees of Indiana University and Indiana
-#                         University Research and Technology
-#                         Corporation.  All rights reserved.
-# Copyright (c) 2004-2005 The University of Tennessee and The University
-#                         of Tennessee Research Foundation.  All rights
-#                         reserved.
-# Copyright (c) 2004-2005 High Performance Computing Center Stuttgart,
-#                         University of Stuttgart.  All rights reserved.
-# Copyright (c) 2004-2005 The Regents of the University of California.
-#                         All rights reserved.
-# Copyright (c) 2011-2013 Los Alamos National Security, LLC.
-#                         All rights reserved.
-# Copyright (c) 2010-2020 Cisco Systems, Inc.  All rights reserved
-# Copyright (c) 2019      Intel, Inc.  All rights reserved.
-# $COPYRIGHT$
-#
-# Additional copyrights may follow
-#
-# $HEADER$
-#
-
-# MCA_oob_tcp_CONFIG([action-if-found], [action-if-not-found])
-# -----------------------------------------------------------
-AC_DEFUN([MCA_prte_oob_tcp_CONFIG],[
-    AC_CONFIG_FILES([src/mca/oob/tcp/Makefile])
-
-    # check for sockaddr_in (a good sign we have TCP)
-    AC_CHECK_TYPES([struct sockaddr_in],
-                   [oob_tcp_happy="yes"],
-                   [oob_tcp_happy="no"],
-                   [AC_INCLUDES_DEFAULT
-#ifdef HAVE_NETINET_IN_H
-#include <netinet/in.h>
-#endif])
-
-    AS_IF([test "$oob_tcp_happy" = "yes"], [$1], [$2])
-])dnl
