+++ src/rml/oob/Makefile.am
@@ -1,0 +1,46 @@
+#
+# Copyright (c) 2004-2005 The Trustees of Indiana University and Indiana
+#                         University Research and Technology
+#                         Corporation.  All rights reserved.
+# Copyright (c) 2004-2005 The University of Tennessee and The University
+#                         of Tennessee Research Foundation.  All rights
+#                         reserved.
+# Copyright (c) 2004-2009 High Performance Computing Center Stuttgart,
+#                         University of Stuttgart.  All rights reserved.
+# Copyright (c) 2004-2005 The Regents of the University of California.
+#                         All rights reserved.
+# Copyright (c) 2010-2020 Cisco Systems, Inc.  All rights reserved
+# Copyright (c) 2012-2013 Los Alamos National Security, LLC.
+#                         All rights reserved
+# Copyright (c) 2014-2020 Intel, Inc.  All rights reserved.
+# Copyright (c) 2017      IBM Corporation.  All rights reserved.
+# Copyright (c) 2022-2024 Nanook Consulting  All rights reserved.
+# $COPYRIGHT$
+#
+# Additional copyrights may follow
+#
+# $HEADER$
+#
+
+dist_prtedata_DATA += \
+      rml/oob/help-oob-base.txt \
+      rml/oob/help-oob-tcp.txt
+
+headers += \
+      rml/oob/oob.h \
+      rml/oob/oob_tcp.h \
+      rml/oob/oob_tcp_listener.h \
+      rml/oob/oob_tcp_common.h \
+      rml/oob/oob_tcp_connection.h \
+      rml/oob/oob_tcp_sendrecv.h \
+      rml/oob/oob_tcp_hdr.h \
+      rml/oob/oob_tcp_peer.h
+
+libprrte_la_SOURCES += \
+      rml/oob/oob_tcp_component.c \
+      rml/oob/oob_tcp.c \
+      rml/oob/oob_tcp_listener.c \
+      rml/oob/oob_tcp_common.c \
+      rml/oob/oob_tcp_connection.c \
+      rml/oob/oob_tcp_sendrecv.c \
+      rml/oob/oob_base_stubs.c
