+++ src/rml/oob/oob_tcp.h
@@ -15,7 +15,7 @@
  * Copyright (c) 2014-2019 Intel, Inc.  All rights reserved.
  * Copyright (c) 2019      Research Organization for Information Science
  *                         and Technology (RIST).  All rights reserved.
- * Copyright (c) 2021-2022 Nanook Consulting.  All rights reserved.
+ * Copyright (c) 2021-2024 Nanook Consulting  All rights reserved.
  * $COPYRIGHT$
  *
  * Additional copyrights may follow
@@ -33,18 +33,13 @@
 #include "src/event/event-internal.h"
 #include "src/mca/base/pmix_base.h"
 
-#include "src/mca/oob/base/base.h"
-#include "src/mca/oob/oob.h"
+#include "src/rml/oob/oob.h"
 
 BEGIN_C_DECLS
 
 /* define some debug levels */
 #define OOB_TCP_DEBUG_FAIL    2
 #define OOB_TCP_DEBUG_CONNECT 7
-
-/* forward declare a couple of structures */
-struct prte_oob_tcp_module_t;
-struct prte_oob_tcp_msg_error_t;
 
 /* define a struct for tracking NIC addresses */
 typedef struct {
@@ -53,19 +48,6 @@ typedef struct {
     struct sockaddr addr;
 } prte_oob_tcp_nicaddr_t;
 PMIX_CLASS_DECLARATION(prte_oob_tcp_nicaddr_t);
-
-/* Module definition */
-typedef void (*prte_oob_tcp_module_accept_connection_fn_t)(const int accepted_fd,
-                                                           const struct sockaddr *addr);
-typedef void (*prte_oob_tcp_module_ping_fn_t)(const pmix_proc_t *proc);
-typedef void (*prte_oob_tcp_module_send_nb_fn_t)(prte_rml_send_t *msg);
-
-typedef struct {
-    prte_oob_tcp_module_accept_connection_fn_t accept_connection;
-    prte_oob_tcp_module_ping_fn_t ping;
-    prte_oob_tcp_module_send_nb_fn_t send_nb;
-} prte_oob_tcp_module_t;
-PRTE_MODULE_EXPORT extern prte_oob_tcp_module_t prte_oob_tcp_module;
 
 /**
  * the state of the connection
@@ -82,10 +64,15 @@ typedef enum {
 } prte_oob_tcp_state_t;
 
 /* module-level shared functions */
-PRTE_MODULE_EXPORT void prte_oob_tcp_send_handler(int fd, short args, void *cbdata);
-PRTE_MODULE_EXPORT void prte_oob_tcp_recv_handler(int fd, short args, void *cbdata);
-PRTE_MODULE_EXPORT void prte_oob_tcp_queue_msg(int sd, short args, void *cbdata);
-
+PRTE_EXPORT void prte_oob_tcp_send_handler(int fd, short args, void *cbdata);
+PRTE_EXPORT void prte_oob_tcp_recv_handler(int fd, short args, void *cbdata);
+PRTE_EXPORT void prte_oob_tcp_queue_msg(int sd, short args, void *cbdata);
+PRTE_EXPORT void prte_oob_accept_connection(const int accepted_fd, const struct sockaddr *addr);
+PRTE_EXPORT void prte_mca_oob_tcp_component_lost_connection(int fd, short args, void *cbdata);
+PRTE_EXPORT void prte_mca_oob_tcp_component_failed_to_connect(int fd, short args, void *cbdata);
+PRTE_EXPORT void prte_mca_oob_tcp_component_no_route(int fd, short args, void *cbdata);
+PRTE_EXPORT void prte_mca_oob_tcp_component_hop_unknown(int fd, short args, void *cbdata);
+PRTE_EXPORT void prte_oob_ping(const pmix_proc_t *proc);
 END_C_DECLS
 
 #endif /* MCA_OOB_TCP_H_ */
