+++ src/rml/oob/oob_tcp_common.c
@@ -16,7 +16,7 @@
  * Copyright (c) 2014-2020 Intel, Inc.  All rights reserved.
  * Copyright (c) 2014-2019 Research Organization for Information Science
  *                         and Technology (RIST).  All rights reserved.
- * Copyright (c) 2021-2022 Nanook Consulting.  All rights reserved.
+ * Copyright (c) 2021-2024 Nanook Consulting  All rights reserved.
  * $COPYRIGHT$
  *
  * Additional copyrights may follow
@@ -63,10 +63,9 @@
 #include "src/util/pmix_net.h"
 #include "src/util/pmix_output.h"
 
-#include "oob_tcp_common.h"
-#include "oob_tcp_peer.h"
-#include "src/mca/oob/tcp/oob_tcp.h"
-#include "src/mca/oob/tcp/oob_tcp_component.h"
+#include "src/rml/oob/oob_tcp_common.h"
+#include "src/rml/oob/oob_tcp_peer.h"
+#include "src/rml/oob/oob_tcp.h"
 
 /**
  * Set socket buffering
@@ -87,27 +86,27 @@ static void set_keepalive(int sd)
     /* Set the option active */
     option = 1;
     if (setsockopt(sd, SOL_SOCKET, SO_KEEPALIVE, &option, optlen) < 0) {
-        pmix_output_verbose(5, prte_oob_base_framework.framework_output,
+        pmix_output_verbose(5, prte_oob_base.output,
                             "[%s:%d] setsockopt(SO_KEEPALIVE) failed: %s (%d)", __FILE__, __LINE__,
                             strerror(prte_socket_errno), prte_socket_errno);
         return;
     }
 #    if defined(TCP_KEEPALIVE)
     /* set the idle time */
-    if (setsockopt(sd, IPPROTO_TCP, TCP_KEEPALIVE, &prte_mca_oob_tcp_component.keepalive_time,
-                   sizeof(prte_mca_oob_tcp_component.keepalive_time))
+    if (setsockopt(sd, IPPROTO_TCP, TCP_KEEPALIVE, &prte_oob_base.keepalive_time,
+                   sizeof(prte_oob_base.keepalive_time))
         < 0) {
-        pmix_output_verbose(5, prte_oob_base_framework.framework_output,
+        pmix_output_verbose(5, prte_oob_base.output,
                             "[%s:%d] setsockopt(TCP_KEEPALIVE) failed: %s (%d)", __FILE__, __LINE__,
                             strerror(prte_socket_errno), prte_socket_errno);
         return;
     }
 #    elif defined(TCP_KEEPIDLE)
     /* set the idle time */
-    if (setsockopt(sd, IPPROTO_TCP, TCP_KEEPIDLE, &prte_mca_oob_tcp_component.keepalive_time,
-                   sizeof(prte_mca_oob_tcp_component.keepalive_time))
+    if (setsockopt(sd, IPPROTO_TCP, TCP_KEEPIDLE, &prte_oob_base.keepalive_time,
+                   sizeof(prte_oob_base.keepalive_time))
         < 0) {
-        pmix_output_verbose(5, prte_oob_base_framework.framework_output,
+        pmix_output_verbose(5, prte_oob_base.output,
                             "[%s:%d] setsockopt(TCP_KEEPIDLE) failed: %s (%d)", __FILE__, __LINE__,
                             strerror(prte_socket_errno), prte_socket_errno);
         return;
@@ -115,10 +114,10 @@ static void set_keepalive(int sd)
 #    endif // TCP_KEEPIDLE
 #    if defined(TCP_KEEPINTVL)
     /* set the keepalive interval */
-    if (setsockopt(sd, IPPROTO_TCP, TCP_KEEPINTVL, &prte_mca_oob_tcp_component.keepalive_intvl,
-                   sizeof(prte_mca_oob_tcp_component.keepalive_intvl))
+    if (setsockopt(sd, IPPROTO_TCP, TCP_KEEPINTVL, &prte_oob_base.keepalive_intvl,
+                   sizeof(prte_oob_base.keepalive_intvl))
         < 0) {
-        pmix_output_verbose(5, prte_oob_base_framework.framework_output,
+        pmix_output_verbose(5, prte_oob_base.output,
                             "[%s:%d] setsockopt(TCP_KEEPINTVL) failed: %s (%d)", __FILE__, __LINE__,
                             strerror(prte_socket_errno), prte_socket_errno);
         return;
@@ -126,10 +125,10 @@ static void set_keepalive(int sd)
 #    endif // TCP_KEEPINTVL
 #    if defined(TCP_KEEPCNT)
     /* set the miss rate */
-    if (setsockopt(sd, IPPROTO_TCP, TCP_KEEPCNT, &prte_mca_oob_tcp_component.keepalive_probes,
-                   sizeof(prte_mca_oob_tcp_component.keepalive_probes))
+    if (setsockopt(sd, IPPROTO_TCP, TCP_KEEPCNT, &prte_oob_base.keepalive_probes,
+                   sizeof(prte_oob_base.keepalive_probes))
         < 0) {
-        pmix_output_verbose(5, prte_oob_base_framework.framework_output,
+        pmix_output_verbose(5, prte_oob_base.output,
                             "[%s:%d] setsockopt(TCP_KEEPCNT) failed: %s (%d)", __FILE__, __LINE__,
                             strerror(prte_socket_errno), prte_socket_errno);
     }
@@ -144,33 +143,33 @@ void prte_oob_tcp_set_socket_options(int sd)
     optval = 1;
     if (setsockopt(sd, IPPROTO_TCP, TCP_NODELAY, (char *) &optval, sizeof(optval)) < 0) {
         prte_backtrace_print(stderr, NULL, 1);
-        pmix_output_verbose(5, prte_oob_base_framework.framework_output,
+        pmix_output_verbose(5, prte_oob_base.output,
                             "[%s:%d] setsockopt(TCP_NODELAY) failed: %s (%d)", __FILE__, __LINE__,
                             strerror(prte_socket_errno), prte_socket_errno);
     }
 #endif
 #if defined(SO_SNDBUF)
-    if (prte_mca_oob_tcp_component.tcp_sndbuf > 0
-        && setsockopt(sd, SOL_SOCKET, SO_SNDBUF, (char *) &prte_mca_oob_tcp_component.tcp_sndbuf,
+    if (prte_oob_base.tcp_sndbuf > 0
+        && setsockopt(sd, SOL_SOCKET, SO_SNDBUF, (char *) &prte_oob_base.tcp_sndbuf,
                       sizeof(int))
                < 0) {
-        pmix_output_verbose(5, prte_oob_base_framework.framework_output,
+        pmix_output_verbose(5, prte_oob_base.output,
                             "[%s:%d] setsockopt(SO_SNDBUF) failed: %s (%d)", __FILE__, __LINE__,
                             strerror(prte_socket_errno), prte_socket_errno);
     }
 #endif
 #if defined(SO_RCVBUF)
-    if (prte_mca_oob_tcp_component.tcp_rcvbuf > 0
-        && setsockopt(sd, SOL_SOCKET, SO_RCVBUF, (char *) &prte_mca_oob_tcp_component.tcp_rcvbuf,
+    if (prte_oob_base.tcp_rcvbuf > 0
+        && setsockopt(sd, SOL_SOCKET, SO_RCVBUF, (char *) &prte_oob_base.tcp_rcvbuf,
                       sizeof(int))
                < 0) {
-        pmix_output_verbose(5, prte_oob_base_framework.framework_output,
+        pmix_output_verbose(5, prte_oob_base.output,
                             "[%s:%d] setsockopt(SO_RCVBUF) failed: %s (%d)", __FILE__, __LINE__,
                             strerror(prte_socket_errno), prte_socket_errno);
     }
 #endif
 
-    if (0 < prte_mca_oob_tcp_component.keepalive_time) {
+    if (0 < prte_oob_base.keepalive_time) {
         set_keepalive(sd);
     }
 }
@@ -179,7 +178,7 @@ prte_oob_tcp_peer_t *prte_oob_tcp_peer_lookup(const pmix_proc_t *name)
 {
     prte_oob_tcp_peer_t *peer;
 
-    PMIX_LIST_FOREACH(peer, &prte_mca_oob_tcp_component.peers, prte_oob_tcp_peer_t)
+    PMIX_LIST_FOREACH(peer, &prte_oob_base.peers, prte_oob_tcp_peer_t)
     {
         if (PMIX_CHECK_PROCID(name, &peer->name)) {
             return peer;
