+++ src/rml/oob/oob_tcp_common.h
@@ -15,7 +15,7 @@
  * Copyright (c) 2014-2019 Intel, Inc.  All rights reserved.
  * Copyright (c) 2019      Research Organization for Information Science
  *                         and Technology (RIST).  All rights reserved.
- * Copyright (c) 2021      Nanook Consulting.  All rights reserved.
+ * Copyright (c) 2021-2024 Nanook Consulting  All rights reserved.
  * $COPYRIGHT$
  *
  * Additional copyrights may follow
@@ -28,10 +28,10 @@
 
 #include "prte_config.h"
 
-#include "oob_tcp.h"
-#include "oob_tcp_peer.h"
+#include "src/rml/oob/oob_tcp.h"
+#include "src/rml/oob/oob_tcp_peer.h"
 
-PRTE_MODULE_EXPORT void prte_oob_tcp_set_socket_options(int sd);
-PRTE_MODULE_EXPORT char *prte_oob_tcp_state_print(prte_oob_tcp_state_t state);
-PRTE_MODULE_EXPORT prte_oob_tcp_peer_t *prte_oob_tcp_peer_lookup(const pmix_proc_t *name);
+PRTE_EXPORT void prte_oob_tcp_set_socket_options(int sd);
+PRTE_EXPORT char *prte_oob_tcp_state_print(prte_oob_tcp_state_t state);
+PRTE_EXPORT prte_oob_tcp_peer_t *prte_oob_tcp_peer_lookup(const pmix_proc_t *name);
 #endif /* _MCA_OOB_TCP_COMMON_H_ */
