+++ src/rml/oob/oob_tcp_connection.h
@@ -15,7 +15,7 @@
  * Copyright (c) 2014-2019 Intel, Inc.  All rights reserved.
  * Copyright (c) 2019      Research Organization for Information Science
  *                         and Technology (RIST).  All rights reserved.
- * Copyright (c) 2021-2023 Nanook Consulting.  All rights reserved.
+ * Copyright (c) 2021-2024 Nanook Consulting  All rights reserved.
  * $COPYRIGHT$
  *
  * Additional copyrights may follow
@@ -35,8 +35,8 @@
 #    include <sys/socket.h>
 #endif
 
-#include "oob_tcp.h"
-#include "oob_tcp_peer.h"
+#include "src/rml/oob/oob_tcp.h"
+#include "src/rml/oob/oob_tcp_peer.h"
 #include "src/threads/pmix_threads.h"
 
 /* State machine for connection operations */
@@ -56,7 +56,7 @@ PMIX_CLASS_DECLARATION(prte_oob_tcp_conn_op_t);
 #define PRTE_ACTIVATE_TCP_CONN_STATE(p, cbfunc)                                             \
     do {                                                                                    \
         prte_oob_tcp_conn_op_t *cop;                                                        \
-        pmix_output_verbose(5, prte_oob_base_framework.framework_output,                    \
+        pmix_output_verbose(5, prte_oob_base.output,                    \
                             "%s:[%s:%d] connect to %s", PRTE_NAME_PRINT(PRTE_PROC_MY_NAME), \
                             __FILE__, __LINE__, PRTE_NAME_PRINT((&(p)->name)));             \
         cop = PMIX_NEW(prte_oob_tcp_conn_op_t);                                             \
@@ -76,7 +76,7 @@ PMIX_CLASS_DECLARATION(prte_oob_tcp_conn_op_t);
 #define PRTE_RETRY_TCP_CONN_STATE(p, cbfunc, tv)                                                  \
     do {                                                                                          \
         prte_oob_tcp_conn_op_t *cop;                                                              \
-        pmix_output_verbose(5, prte_oob_base_framework.framework_output,                          \
+        pmix_output_verbose(5, prte_oob_base.output,                          \
                             "%s:[%s:%d] retry connect to %s", PRTE_NAME_PRINT(PRTE_PROC_MY_NAME), \
                             __FILE__, __LINE__, PRTE_NAME_PRINT((&(p)->name)));                   \
         cop = PMIX_NEW(prte_oob_tcp_conn_op_t);                                                   \
@@ -86,12 +86,12 @@ PMIX_CLASS_DECLARATION(prte_oob_tcp_conn_op_t);
         prte_event_evtimer_add(&cop->ev, (tv));                                                   \
     } while (0);
 
-PRTE_MODULE_EXPORT void prte_oob_tcp_peer_try_connect(int fd, short args, void *cbdata);
-PRTE_MODULE_EXPORT void prte_oob_tcp_peer_dump(prte_oob_tcp_peer_t *peer, const char *msg);
-PRTE_MODULE_EXPORT bool prte_oob_tcp_peer_accept(prte_oob_tcp_peer_t *peer);
-PRTE_MODULE_EXPORT void prte_oob_tcp_peer_complete_connect(prte_oob_tcp_peer_t *peer);
-PRTE_MODULE_EXPORT int prte_oob_tcp_peer_recv_connect_ack(prte_oob_tcp_peer_t *peer, int sd,
+PRTE_EXPORT void prte_oob_tcp_peer_try_connect(int fd, short args, void *cbdata);
+PRTE_EXPORT void prte_oob_tcp_peer_dump(prte_oob_tcp_peer_t *peer, const char *msg);
+PRTE_EXPORT bool prte_oob_tcp_peer_accept(prte_oob_tcp_peer_t *peer);
+PRTE_EXPORT void prte_oob_tcp_peer_complete_connect(prte_oob_tcp_peer_t *peer);
+PRTE_EXPORT int prte_oob_tcp_peer_recv_connect_ack(prte_oob_tcp_peer_t *peer, int sd,
                                                           prte_oob_tcp_hdr_t *dhdr);
-PRTE_MODULE_EXPORT void prte_oob_tcp_peer_close(prte_oob_tcp_peer_t *peer);
+PRTE_EXPORT void prte_oob_tcp_peer_close(prte_oob_tcp_peer_t *peer);
 
 #endif /* _MCA_OOB_TCP_CONNECTION_H_ */
