+++ src/rml/oob/oob_tcp_listener.c
@@ -68,12 +68,11 @@
 #include "src/util/pmix_parse_options.h"
 #include "src/util/pmix_show_help.h"
 
-#include "src/mca/oob/tcp/oob_tcp.h"
-#include "src/mca/oob/tcp/oob_tcp_common.h"
-#include "src/mca/oob/tcp/oob_tcp_component.h"
-#include "src/mca/oob/tcp/oob_tcp_connection.h"
-#include "src/mca/oob/tcp/oob_tcp_listener.h"
-#include "src/mca/oob/tcp/oob_tcp_peer.h"
+#include "src/rml/oob/oob_tcp.h"
+#include "src/rml/oob/oob_tcp_common.h"
+#include "src/rml/oob/oob_tcp_connection.h"
+#include "src/rml/oob/oob_tcp_listener.h"
+#include "src/rml/oob/oob_tcp_peer.h"
 
 static void connection_event_handler(int incoming_sd, short flags, void *cbdata);
 static void *listen_thread(pmix_object_t *obj);
@@ -101,9 +100,9 @@ int prte_oob_tcp_start_listening(void)
     prte_oob_tcp_listener_t *listener;
 
     /* if we don't have any TCP interfaces, we shouldn't be here */
-    if (NULL == prte_mca_oob_tcp_component.ipv4conns
+    if (NULL == prte_oob_base.ipv4conns
 #if PRTE_ENABLE_IPV6
-        && NULL == prte_mca_oob_tcp_component.ipv6conns
+        && NULL == prte_oob_base.ipv6conns
 #endif
     ) {
         PRTE_ERROR_LOG(PRTE_ERR_NOT_FOUND);
@@ -128,25 +127,25 @@ int prte_oob_tcp_start_listening(void)
      * harvest connection requests as rapidly as possible
      */
     if (PRTE_PROC_IS_MASTER) {
-        if (0 > pipe(prte_mca_oob_tcp_component.stop_thread)) {
+        if (0 > pipe(prte_oob_base.stop_thread)) {
             PRTE_ERROR_LOG(PRTE_ERR_OUT_OF_RESOURCE);
             return PRTE_ERR_OUT_OF_RESOURCE;
         }
 
         /* Make sure the pipe FDs are set to close-on-exec so that
            they don't leak into children */
-        if (pmix_fd_set_cloexec(prte_mca_oob_tcp_component.stop_thread[0]) != PRTE_SUCCESS
-            || pmix_fd_set_cloexec(prte_mca_oob_tcp_component.stop_thread[1]) != PRTE_SUCCESS) {
-            close(prte_mca_oob_tcp_component.stop_thread[0]);
-            close(prte_mca_oob_tcp_component.stop_thread[1]);
+        if (pmix_fd_set_cloexec(prte_oob_base.stop_thread[0]) != PRTE_SUCCESS
+            || pmix_fd_set_cloexec(prte_oob_base.stop_thread[1]) != PRTE_SUCCESS) {
+            close(prte_oob_base.stop_thread[0]);
+            close(prte_oob_base.stop_thread[1]);
             PRTE_ERROR_LOG(PRTE_ERR_IN_ERRNO);
             return PRTE_ERR_IN_ERRNO;
         }
 
-        prte_mca_oob_tcp_component.listen_thread_active = true;
-        prte_mca_oob_tcp_component.listen_thread.t_run = listen_thread;
-        prte_mca_oob_tcp_component.listen_thread.t_arg = NULL;
-        if (PRTE_SUCCESS != (rc = pmix_thread_start(&prte_mca_oob_tcp_component.listen_thread))) {
+        prte_oob_base.listen_thread_active = true;
+        prte_oob_base.listen_thread.t_run = listen_thread;
+        prte_oob_base.listen_thread.t_arg = NULL;
+        if (PRTE_SUCCESS != (rc = pmix_thread_start(&prte_oob_base.listen_thread))) {
             PRTE_ERROR_LOG(rc);
             pmix_output(0, "%s Unable to start listen thread", PRTE_NAME_PRINT(PRTE_PROC_MY_NAME));
         }
@@ -155,7 +154,7 @@ int prte_oob_tcp_start_listening(void)
 
     /* otherwise, setup to listen via the event lib */
 
-    PMIX_LIST_FOREACH(listener, &prte_mca_oob_tcp_component.listeners, prte_oob_tcp_listener_t)
+    PMIX_LIST_FOREACH(listener, &prte_oob_base.listeners, prte_oob_tcp_listener_t)
     {
         listener->ev_active = true;
         prte_event_set(prte_event_base, &listener->event, listener->sd,
@@ -191,16 +190,16 @@ static int create_listen(void)
      * port in the range.  Otherwise, tcp_port_min will be 0, which
      * means "pick any port"
      */
-    if (NULL != prte_mca_oob_tcp_component.tcp_static_ports) {
+    if (NULL != prte_oob_base.tcp_static_ports) {
         /* if static ports were provided, take the
          * first entry in the list
          */
-        PMIX_ARGV_APPEND_NOSIZE_COMPAT(&ports, prte_mca_oob_tcp_component.tcp_static_ports[0]);
+        PMIX_ARGV_APPEND_NOSIZE_COMPAT(&ports, prte_oob_base.tcp_static_ports[0]);
         /* flag that we are using static ports */
         prte_static_ports = true;
-    } else if (NULL != prte_mca_oob_tcp_component.tcp_dyn_ports) {
+    } else if (NULL != prte_oob_base.tcp_dyn_ports) {
         /* take the entire range */
-        ports = PMIX_ARGV_COPY_COMPAT(prte_mca_oob_tcp_component.tcp_dyn_ports);
+        ports = PMIX_ARGV_COPY_COMPAT(prte_oob_base.tcp_dyn_ports);
         prte_static_ports = false;
     } else {
         /* flag the system to dynamically take any available port */
@@ -225,7 +224,7 @@ static int create_listen(void)
      * sockets to support more flexible wireup protocols
      */
     for (i = 0; i < PMIX_ARGV_COUNT_COMPAT(ports); i++) {
-        pmix_output_verbose(5, prte_oob_base_framework.framework_output,
+        pmix_output_verbose(5, prte_oob_base.output,
                             "%s attempting to bind to IPv4 port %s",
                             PRTE_NAME_PRINT(PRTE_PROC_MY_NAME), ports[i]);
         /* get the port number */
@@ -239,7 +238,7 @@ static int create_listen(void)
         sd = socket(AF_INET, SOCK_STREAM, 0);
         if (sd < 0) {
             if (EAFNOSUPPORT != prte_socket_errno) {
-                pmix_output(0, "prte_mca_oob_tcp_component_init: socket() failed: %s (%d)",
+                pmix_output(0, "prte_oob_create_listen: socket() failed: %s (%d)",
                             strerror(prte_socket_errno), prte_socket_errno);
             }
             PMIX_ARGV_FREE_COMPAT(ports);
@@ -254,7 +253,7 @@ static int create_listen(void)
         }
         if (setsockopt(sd, SOL_SOCKET, SO_REUSEADDR, (const char *) &flags, sizeof(flags)) < 0) {
             pmix_output(0,
-                        "prte_oob_tcp_create_listen: unable to set the "
+                        "prte_oob_create_listen: unable to set the "
                         "SO_REUSEADDR option (%s:%d)\n",
                         strerror(prte_socket_errno), prte_socket_errno);
             CLOSE_THE_SOCKET(sd);
@@ -266,7 +265,7 @@ static int create_listen(void)
            this FD */
         if (pmix_fd_set_cloexec(sd) != PRTE_SUCCESS) {
             pmix_output(0,
-                        "prte_oob_tcp_create_listen: unable to set the "
+                        "prte_oob_create_listen: unable to set the "
                         "listening socket to CLOEXEC (%s:%d)\n",
                         strerror(prte_socket_errno), prte_socket_errno);
             CLOSE_THE_SOCKET(sd);
@@ -287,7 +286,7 @@ static int create_listen(void)
         }
         /* resolve assigned port */
         if (getsockname(sd, (struct sockaddr *) &inaddr, &addrlen) < 0) {
-            pmix_output(0, "prte_oob_tcp_create_listen: getsockname(): %s (%d)",
+            pmix_output(0, "prte_oob_create_listen: getsockname(): %s (%d)",
                         strerror(prte_socket_errno), prte_socket_errno);
             CLOSE_THE_SOCKET(sd);
             PMIX_ARGV_FREE_COMPAT(ports);
@@ -296,7 +295,7 @@ static int create_listen(void)
 
         /* setup listen backlog to maximum allowed by kernel */
         if (listen(sd, SOMAXCONN) < 0) {
-            pmix_output(0, "prte_mca_oob_tcp_component_init: listen(): %s (%d)",
+            pmix_output(0, "prte_oob_create_listen: listen(): %s (%d)",
                         strerror(prte_socket_errno), prte_socket_errno);
             CLOSE_THE_SOCKET(sd);
             PMIX_ARGV_FREE_COMPAT(ports);
@@ -305,7 +304,7 @@ static int create_listen(void)
 
         /* set socket up to be non-blocking, otherwise accept could block */
         if ((flags = fcntl(sd, F_GETFL, 0)) < 0) {
-            pmix_output(0, "prte_mca_oob_tcp_component_init: fcntl(F_GETFL) failed: %s (%d)",
+            pmix_output(0, "prte_oob_create_listen init: fcntl(F_GETFL) failed: %s (%d)",
                         strerror(prte_socket_errno), prte_socket_errno);
             CLOSE_THE_SOCKET(sd);
             PMIX_ARGV_FREE_COMPAT(ports);
@@ -313,7 +312,7 @@ static int create_listen(void)
         }
         flags |= O_NONBLOCK;
         if (fcntl(sd, F_SETFL, flags) < 0) {
-            pmix_output(0, "prte_mca_oob_tcp_component_init: fcntl(F_SETFL) failed: %s (%d)",
+            pmix_output(0, "prte_oob_create_listen init: fcntl(F_SETFL) failed: %s (%d)",
                         strerror(prte_socket_errno), prte_socket_errno);
             CLOSE_THE_SOCKET(sd);
             PMIX_ARGV_FREE_COMPAT(ports);
@@ -328,13 +327,13 @@ static int create_listen(void)
             /* save the first one */
             prte_process_info.my_port = conn->port;
         }
-        pmix_list_append(&prte_mca_oob_tcp_component.listeners, &conn->item);
+        pmix_list_append(&prte_oob_base.listeners, &conn->item);
         /* and to our ports */
         pmix_asprintf(&tconn, "%d", ntohs(((struct sockaddr_in *) &inaddr)->sin_port));
-        PMIX_ARGV_APPEND_NOSIZE_COMPAT(&prte_mca_oob_tcp_component.ipv4ports, tconn);
+        PMIX_ARGV_APPEND_NOSIZE_COMPAT(&prte_oob_base.ipv4ports, tconn);
         free(tconn);
         if (OOB_TCP_DEBUG_CONNECT
-            <= pmix_output_get_verbosity(prte_oob_base_framework.framework_output)) {
+            <= pmix_output_get_verbosity(prte_oob_base.output)) {
             port = ntohs(((struct sockaddr_in *) &inaddr)->sin_port);
             pmix_output(0, "%s assigned IPv4 port %d", PRTE_NAME_PRINT(PRTE_PROC_MY_NAME), port);
         }
@@ -347,7 +346,7 @@ static int create_listen(void)
     /* done with this, so release it */
     PMIX_ARGV_FREE_COMPAT(ports);
 
-    if (0 == pmix_list_get_size(&prte_mca_oob_tcp_component.listeners)) {
+    if (0 == pmix_list_get_size(&prte_oob_base.listeners)) {
         /* cleanup */
         if (0 <= sd) {
             CLOSE_THE_SOCKET(sd);
@@ -384,16 +383,16 @@ static int create_listen6(void)
      * means "pick any port"
      */
     if (PRTE_PROC_IS_DAEMON) {
-        if (NULL != prte_mca_oob_tcp_component.tcp6_static_ports) {
+        if (NULL != prte_oob_base.tcp6_static_ports) {
             /* if static ports were provided, take the
              * first entry in the list
              */
-            PMIX_ARGV_APPEND_NOSIZE_COMPAT(&ports, prte_mca_oob_tcp_component.tcp6_static_ports[0]);
+            PMIX_ARGV_APPEND_NOSIZE_COMPAT(&ports, prte_oob_base.tcp6_static_ports[0]);
             /* flag that we are using static ports */
             prte_static_ports = true;
-        } else if (NULL != prte_mca_oob_tcp_component.tcp6_dyn_ports) {
+        } else if (NULL != prte_oob_base.tcp6_dyn_ports) {
             /* take the entire range */
-            ports = PMIX_ARGV_COPY_COMPAT(prte_mca_oob_tcp_component.tcp6_dyn_ports);
+            ports = PMIX_ARGV_COPY_COMPAT(prte_oob_base.tcp6_dyn_ports);
             prte_static_ports = false;
         } else {
             /* flag the system to dynamically take any available port */
@@ -401,16 +400,16 @@ static int create_listen6(void)
             prte_static_ports = false;
         }
     } else {
-        if (NULL != prte_mca_oob_tcp_component.tcp6_static_ports) {
+        if (NULL != prte_oob_base.tcp6_static_ports) {
             /* if static ports were provided, take the
              * first entry in the list
              */
-            PMIX_ARGV_APPEND_NOSIZE_COMPAT(&ports, prte_mca_oob_tcp_component.tcp6_static_ports[0]);
+            PMIX_ARGV_APPEND_NOSIZE_COMPAT(&ports, prte_oob_base.tcp6_static_ports[0]);
             /* flag that we are using static ports */
             prte_static_ports = true;
-        } else if (NULL != prte_mca_oob_tcp_component.tcp6_dyn_ports) {
+        } else if (NULL != prte_oob_base.tcp6_dyn_ports) {
             /* take the entire range */
-            ports = PMIX_ARGV_COPY_COMPAT(prte_mca_oob_tcp_component.tcp6_dyn_ports);
+            ports = PMIX_ARGV_COPY_COMPAT(prte_oob_base.tcp6_dyn_ports);
             prte_static_ports = false;
         } else {
             /* flag the system to dynamically take any available port */
@@ -436,7 +435,7 @@ static int create_listen6(void)
      * sockets to support more flexible wireup protocols
      */
     for (i = 0; i < PMIX_ARGV_COUNT_COMPAT(ports); i++) {
-        pmix_output_verbose(5, prte_oob_base_framework.framework_output,
+        pmix_output_verbose(5, prte_oob_base.output,
                             "%s attempting to bind to IPv6 port %s",
                             PRTE_NAME_PRINT(PRTE_PROC_MY_NAME), ports[i]);
         /* get the port number */
@@ -450,7 +449,7 @@ static int create_listen6(void)
         sd = socket(AF_INET6, SOCK_STREAM, 0);
         if (sd < 0) {
             if (EAFNOSUPPORT != prte_socket_errno) {
-                pmix_output(0, "prte_mca_oob_tcp_component_init: socket() failed: %s (%d)",
+                pmix_output(0, "prte_oob_create_listen6: socket() failed: %s (%d)",
                             strerror(prte_socket_errno), prte_socket_errno);
             }
             return PRTE_ERR_IN_ERRNO;
@@ -459,7 +458,7 @@ static int create_listen6(void)
            this FD */
         if (pmix_fd_set_cloexec(sd) != PRTE_SUCCESS) {
             pmix_output(0,
-                        "prte_oob_tcp_create_listen6: unable to set the "
+                        "prte_oob_create_listen6: unable to set the "
                         "listening socket to CLOEXEC (%s:%d)\n",
                         strerror(prte_socket_errno), prte_socket_errno);
             CLOSE_THE_SOCKET(sd);
@@ -496,7 +495,7 @@ static int create_listen6(void)
         }
         /* resolve assigned port */
         if (getsockname(sd, (struct sockaddr *) &inaddr, &addrlen) < 0) {
-            pmix_output(0, "prte_oob_tcp_create_listen: getsockname(): %s (%d)",
+            pmix_output(0, "prte_oob_create_listen6: getsockname(): %s (%d)",
                         strerror(prte_socket_errno), prte_socket_errno);
             CLOSE_THE_SOCKET(sd);
             return PRTE_ERROR;
@@ -504,20 +503,20 @@ static int create_listen6(void)
 
         /* setup listen backlog to maximum allowed by kernel */
         if (listen(sd, SOMAXCONN) < 0) {
-            pmix_output(0, "prte_mca_oob_tcp_component_init: listen(): %s (%d)",
+            pmix_output(0, "prte_oob_create_listen6: listen(): %s (%d)",
                         strerror(prte_socket_errno), prte_socket_errno);
             return PRTE_ERROR;
         }
 
         /* set socket up to be non-blocking, otherwise accept could block */
         if ((flags = fcntl(sd, F_GETFL, 0)) < 0) {
-            pmix_output(0, "prte_mca_oob_tcp_component_init: fcntl(F_GETFL) failed: %s (%d)",
+            pmix_output(0, "prte_oob_create_listen6: fcntl(F_GETFL) failed: %s (%d)",
                         strerror(prte_socket_errno), prte_socket_errno);
             return PRTE_ERROR;
         }
         flags |= O_NONBLOCK;
         if (fcntl(sd, F_SETFL, flags) < 0) {
-            pmix_output(0, "prte_mca_oob_tcp_component_init: fcntl(F_SETFL) failed: %s (%d)",
+            pmix_output(0, "prte_oob_create_listen6: fcntl(F_SETFL) failed: %s (%d)",
                         strerror(prte_socket_errno), prte_socket_errno);
             return PRTE_ERROR;
         }
@@ -527,13 +526,13 @@ static int create_listen6(void)
         conn->tcp6 = true;
         conn->sd = sd;
         conn->port = ntohs(((struct sockaddr_in6 *) &inaddr)->sin6_port);
-        pmix_list_append(&prte_mca_oob_tcp_component.listeners, &conn->item);
+        pmix_list_append(&prte_oob_base.listeners, &conn->item);
         /* and to our ports */
         pmix_asprintf(&tconn, "%d", ntohs(((struct sockaddr_in6 *) &inaddr)->sin6_port));
-        PMIX_ARGV_APPEND_NOSIZE_COMPAT(&prte_mca_oob_tcp_component.ipv6ports, tconn);
+        PMIX_ARGV_APPEND_NOSIZE_COMPAT(&prte_oob_base.ipv6ports, tconn);
         free(tconn);
         if (OOB_TCP_DEBUG_CONNECT
-            <= pmix_output_get_verbosity(prte_oob_base_framework.framework_output)) {
+            <= pmix_output_get_verbosity(prte_oob_base.output)) {
             pmix_output(0, "%s assigned IPv6 port %d", PRTE_NAME_PRINT(PRTE_PROC_MY_NAME),
                         (int) ntohs(((struct sockaddr_in6 *) &inaddr)->sin6_port));
         }
@@ -543,7 +542,7 @@ static int create_listen6(void)
             break;
         }
     }
-    if (0 == pmix_list_get_size(&prte_mca_oob_tcp_component.listeners)) {
+    if (0 == pmix_list_get_size(&prte_oob_base.listeners)) {
         /* cleanup */
         CLOSE_THE_SOCKET(sd);
         PMIX_ARGV_FREE_COMPAT(ports);
@@ -579,28 +578,27 @@ static void *listen_thread(pmix_object_t *obj)
      * to the event method for handling any further connections
      * so as to minimize overhead
      */
-    while (prte_mca_oob_tcp_component.listen_thread_active) {
+    while (prte_oob_base.listen_thread_active) {
         FD_ZERO(&readfds);
         max = -1;
-        PMIX_LIST_FOREACH(listener, &prte_mca_oob_tcp_component.listeners, prte_oob_tcp_listener_t)
+        PMIX_LIST_FOREACH(listener, &prte_oob_base.listeners, prte_oob_tcp_listener_t)
         {
             FD_SET(listener->sd, &readfds);
             max = (listener->sd > max) ? listener->sd : max;
         }
         /* add the stop_thread fd */
-        FD_SET(prte_mca_oob_tcp_component.stop_thread[0], &readfds);
-        max = (prte_mca_oob_tcp_component.stop_thread[0] > max) ? prte_mca_oob_tcp_component.stop_thread[0]
-                                                            : max;
+        FD_SET(prte_oob_base.stop_thread[0], &readfds);
+        max = (prte_oob_base.stop_thread[0] > max) ? prte_oob_base.stop_thread[0] : max;
 
         /* set timeout interval */
-        timeout.tv_sec = prte_mca_oob_tcp_component.listen_thread_tv.tv_sec;
-        timeout.tv_usec = prte_mca_oob_tcp_component.listen_thread_tv.tv_usec;
+        timeout.tv_sec = prte_oob_base.listen_thread_tv.tv_sec;
+        timeout.tv_usec = prte_oob_base.listen_thread_tv.tv_usec;
 
         /* Block in a select to avoid hammering the cpu.  If a connection
          * comes in, we'll get woken up right away.
          */
         rc = select(max + 1, &readfds, NULL, NULL, &timeout);
-        if (!prte_mca_oob_tcp_component.listen_thread_active) {
+        if (!prte_oob_base.listen_thread_active) {
             /* we've been asked to terminate */
             return NULL;
         }
@@ -617,7 +615,7 @@ static void *listen_thread(pmix_object_t *obj)
          */
         do {
             accepted_connections = 0;
-            PMIX_LIST_FOREACH(listener, &prte_mca_oob_tcp_component.listeners, prte_oob_tcp_listener_t)
+            PMIX_LIST_FOREACH(listener, &prte_oob_base.listeners, prte_oob_tcp_listener_t)
             {
                 sd = listener->sd;
 
@@ -677,7 +675,7 @@ static void *listen_thread(pmix_object_t *obj)
                     }
                 }
 
-                pmix_output_verbose(OOB_TCP_DEBUG_CONNECT, prte_oob_base_framework.framework_output,
+                pmix_output_verbose(OOB_TCP_DEBUG_CONNECT, prte_oob_base.output,
                                     "%s prte_oob_tcp_listen_thread: incoming connection: "
                                     "(%d, %d) %s:%d\n",
                                     PRTE_NAME_PRINT(PRTE_PROC_MY_NAME), pending_connection->fd,
@@ -716,28 +714,6 @@ static void *listen_thread(pmix_object_t *obj)
     }
 
 done:
-#if 0
-    /* once we complete the initial launch, the "flood" of connections
-     * will end - only connection requests from local procs, connect/accept
-     * operations across mpirun instances, or the occasional tool will need
-     * to be serviced. As these are relatively small events, we can easily
-     * handle them in the context of the event library and no longer require
-     * a separate connection harvesting thread. So switch over to the event
-     * lib handler now
-     */
-    pmix_output_verbose(OOB_TCP_DEBUG_CONNECT, prte_oob_base_framework.framework_output,
-                        "%s prte_oob_tcp_listen_thread: switching to event lib",
-                        PRTE_NAME_PRINT(PRTE_PROC_MY_NAME));
-    /* setup to listen via event library */
-    PMIX_LIST_FOREACH(listener, &prte_mca_oob_tcp_component.listeners, prte_oob_tcp_listener_t) {
-        prte_event_set(prte_event_base, listener->event,
-                   listener->sd,
-                   PRTE_EV_READ|PRTE_EV_PERSIST,
-                   connection_event_handler,
-                   0);
-        prte_event_add(listener->event, 0);
-    }
-#endif
     return NULL;
 }
 
@@ -753,7 +729,7 @@ static void connection_handler(int sd, short flags, void *cbdata)
 
     PMIX_ACQUIRE_OBJECT(new_connection);
 
-    pmix_output_verbose(4, prte_oob_base_framework.framework_output,
+    pmix_output_verbose(4, prte_oob_base.output,
                         "%s connection_handler: working connection "
                         "(%d, %d) %s:%d\n",
                         PRTE_NAME_PRINT(PRTE_PROC_MY_NAME), new_connection->fd, prte_socket_errno,
@@ -761,8 +737,8 @@ static void connection_handler(int sd, short flags, void *cbdata)
                         pmix_net_get_port((struct sockaddr *) &new_connection->addr));
 
     /* process the connection */
-    prte_oob_tcp_module.accept_connection(new_connection->fd,
-                                          (struct sockaddr *) &(new_connection->addr));
+    prte_oob_accept_connection(new_connection->fd, (struct sockaddr *) &(new_connection->addr));
+
     /* cleanup */
     PMIX_RELEASE(new_connection);
 }
@@ -778,7 +754,7 @@ static void connection_event_handler(int incoming_sd, short flags, void *cbdata)
     PRTE_HIDE_UNUSED_PARAMS(flags, cbdata);
 
     sd = accept(incoming_sd, (struct sockaddr *) &addr, &addrlen);
-    pmix_output_verbose(OOB_TCP_DEBUG_CONNECT, prte_oob_base_framework.framework_output,
+    pmix_output_verbose(OOB_TCP_DEBUG_CONNECT, prte_oob_base.output,
                         "%s connection_event_handler: working connection "
                         "(%d, %d) %s:%d\n",
                         PRTE_NAME_PRINT(PRTE_PROC_MY_NAME), sd, prte_socket_errno,
@@ -815,7 +791,7 @@ static void connection_event_handler(int incoming_sd, short flags, void *cbdata)
     }
 
     /* process the connection */
-    prte_oob_tcp_module.accept_connection(sd, &addr);
+    prte_oob_accept_connection(sd, &addr);
 }
 
 static void tcp_ev_cons(prte_oob_tcp_listener_t *event)
